+++ movie/src/main/java/pl/pjatk/movie/exceptions/MovieNotFoundException.java
@@ -1,7 +1,7 @@
 package pl.pjatk.movie.exceptions;
 
 public class MovieNotFoundException extends RuntimeException{
-    public MovieNotFoundException(int id) {
+    public MovieNotFoundException(Long id) {
         super("Film o ID " + id + " nie istnieje");
     }
 }
